users but role

+++ db/migrate/20200415112850_devise_create_admins.rb
@@ -36,7 +36,7 @@ class DeviseCreateAdmins < ActiveRecord::Migration[6.0]
       t.date :dob
       t.integer :mobile
       t.string :adress
-      t.integer :NationalId
+      t.bigint :NationalId
       t.integer :Role
       t.string :gender
